requirements.txt
# Dependências do Framework Multi-Paradigma
# Autor: Frederico
# Data: Fevereiro de 2025

# Manipulação de dados
pandas>=1.5.0
numpy>=1.23.0

# Machine Learning
scikit-learn>=1.2.0
xgboost>=1.7.0
lightgbm>=3.3.0
catboost>=1.1.0
imbalanced-learn>=0.10.0

# Explainable AI
shap>=0.41.0
lime>=0.2.0

# Visualização
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.11.0

# Configuração
pyyaml>=6.0

# Utilitários
joblib>=1.2.0
tqdm>=4.64.0
